pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mdfileconvert"
version = "2.0.0"
authors = [
  { name = "Leandro Rivera Ríos", email = "leo.232rivera@gmail.com" },
]
description = "Convierte archivos a formato Markdown extrayendo texto, tablas e imágenes."
readme = "README.md"
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pdfplumber",             
  "Pillow",                 
  "python-docx",            
  "pandas",                 
  "openpyxl",               
  "python-pptx",            
  "lxml",     
  "tabulate",
  "numpy",
  "python-dateutil",              
]

[project.urls]
"Homepage" = "https://github.com/LeoR22/mdfileconvert"
